SINGA-84 Header Files Rearrange
Move deprecated input layers into input/deprecated.h and input/deprecated.cc
Add "deprecated" tag in the comments.
Remove the DataShard class in utils/ folder, use it as an alias of io::KVFile.
