+++ include/singa/neuralnet/connection_layer/concate.h
@@ -25,10 +25,6 @@
 #include <vector>
 #include "singa/neuralnet/layer.h"
 
-/**
- * \file this file includes the declarations of layers that inherit the
- * base ConnectionLayer.
- */
 namespace singa {
 /**
  * Connect multiple (src) layers with a single (dst) layer.
