+++ include/singa/neuralnet/neuron_layer/inner_product.h
@@ -24,12 +24,6 @@
 
 #include <vector>
 #include "singa/neuralnet/layer.h"
-#include "singa/proto/job.pb.h"
-
-/**
- * \file this file includes the declarations neuron layer classes that conduct
- * the transformation of features.
- */
 namespace singa {
 
 class InnerProductLayer : public NeuronLayer {
