+++ include/singa/neuralnet/neuron_layer/pooling.h
@@ -26,10 +26,6 @@
 #include "singa/neuralnet/layer.h"
 #include "singa/proto/job.pb.h"
 
-/**
- * \file this file includes the declarations neuron layer classes that conduct
- * the transformation of features.
- */
 namespace singa {
 
 class PoolingLayer : public NeuronLayer {
