+++ include/singa/neuralnet/neuron_layer/rbm.h
@@ -26,10 +26,6 @@
 #include "singa/neuralnet/layer.h"
 #include "singa/proto/job.pb.h"
 
-/**
- * \file this file includes the declarations neuron layer classes that conduct
- * the transformation of features.
- */
 namespace singa {
 /**
  * Base layer for RBM models.
