+++ include/singa/neuralnet/neuron_layer/relu.h
@@ -25,11 +25,6 @@
 #include <vector>
 #include "singa/neuralnet/layer.h"
 #include "singa/proto/job.pb.h"
-
-/**
- * \file this file includes the declarations neuron layer classes that conduct
- * the transformation of features.
- */
 namespace singa {
 
 class ReLULayer : public NeuronLayer {
