+++ src/driver.cc
@@ -37,17 +37,11 @@
 #include "singa/neuralnet/connection_layer/concate.h"
 #include "singa/neuralnet/connection_layer/slice.h"
 #include "singa/neuralnet/connection_layer/split.h"
-#include "singa/neuralnet/input_layer/parser.h"
+#include "singa/neuralnet/input_layer/deprecated.h"
 #include "singa/neuralnet/input_layer/csv_record.h"
-#include "singa/neuralnet/input_layer/data.h"
 #include "singa/neuralnet/input_layer/image_preprocess.h"
-#include "singa/neuralnet/input_layer/label.h"
-#include "singa/neuralnet/input_layer/lmdb_data.h"
-#include "singa/neuralnet/input_layer/mnist.h"
 #include "singa/neuralnet/input_layer/prefetch.h"
 #include "singa/neuralnet/input_layer/proto_record.h"
-#include "singa/neuralnet/input_layer/rgb_image.h"
-#include "singa/neuralnet/input_layer/shard_data.h"
 #include "singa/neuralnet/input_layer/store_input.h"
 #include "singa/neuralnet/loss_layer/euclidean.h"
 #include "singa/neuralnet/loss_layer/softmax.h"
@@ -61,6 +55,7 @@
 #include "singa/neuralnet/neuron_layer/sigmoid.h"
 #include "singa/neuralnet/neuron_layer/stanh.h"
 #include "singa/neuralnet/output_layer/output_layer.h"
+
 
 extern "C" void openblas_set_num_threads(int num);
